feat: legacy MVC 삭제
