--- app/src/main/java/com/techcourse/controller/AnnotationLoginController.java
@@ -1,52 +1,0 @@
-package com.techcourse.controller;
-
-import com.techcourse.domain.User;
-import com.techcourse.repository.InMemoryUserRepository;
-import context.org.springframework.stereotype.Controller;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import web.org.springframework.web.bind.annotation.RequestMapping;
-import web.org.springframework.web.bind.annotation.RequestMethod;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.view.JspView;
-
-@Controller
-public class AnnotationLoginController {
-
-    private static final Logger log = LoggerFactory.getLogger(AnnotationLoginController.class);
-
-    @RequestMapping(value = "/ano/login", method = RequestMethod.POST)
-    public ModelAndView postLogin(HttpServletRequest req, HttpServletResponse res) {
-        if (UserSession.isLoggedIn(req.getSession())) {
-            return new ModelAndView(new JspView("redirect:/index.jsp"));
-        }
-
-        return InMemoryUserRepository.findByAccount(req.getParameter("account"))
-                .map(user -> {
-                    log.info("User : {}", user);
-                    return login(req, user);
-                })
-                .orElse(new ModelAndView(new JspView("redirect:/401.jsp")));
-    }
-
-    private ModelAndView login(final HttpServletRequest request, final User user) {
-        if (user.checkPassword(request.getParameter("password"))) {
-            final var session = request.getSession();
-            session.setAttribute(UserSession.SESSION_KEY, user);
-            return new ModelAndView(new JspView("redirect:/index.jsp"));
-        }
-        return new ModelAndView(new JspView("redirect:/401.jsp"));
-    }
-
-    @RequestMapping(value = "/ano/login", method = RequestMethod.GET)
-    public ModelAndView getLogin(HttpServletRequest req, HttpServletResponse res) {
-        return UserSession.getUserFrom(req.getSession())
-                .map(user -> {
-                    log.info("logged in {}", user.getAccount());
-                    return new ModelAndView(new JspView("redirect:/index.jsp"));
-                })
-                .orElse(new ModelAndView(new JspView("/login.jsp")));
-    }
-}
