--- mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/handleradapter/ControllerHandlerAdapter.java
@@ -1,24 +1,0 @@
-package webmvc.org.springframework.web.servlet.mvc.handleradapter;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.view.JspView;
-
-import javax.annotation.Nullable;
-
-public class ControllerHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(Object handler) {
-        return (handler instanceof Controller);
-    }
-
-    @Override
-    @Nullable
-    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        String viewName = ((Controller) handler).execute(request, response);
-        return new ModelAndView(new JspView(viewName));
-    }
-}
